Exec rq in entrypoint
Make the entrypoint scripts exec rq at the end so sending a TERM to the
container entrypoint will get that TERM delivered to the rq process
cleanly. Normally, bash eats the TERM and leaves rq running as an
orphan. Then when the container is cleaned up, the processes are KILL'd,
which does not result in a clean exit.

Added the exec to the guniocrn in content-server and api. Unclean exit
isn't as big of a deal with those, but since gunicorn also handles TERM
and exits cleanly when sessions are done, it seems like a sane idea for
them too.

I think tini can be left out as well, but we'll leave that for another
day.