Stop using hashFiles function
The actions `hashFiles` function does not work with the env context, so throw that garbage on the floor and just use sha256sum with a shell var.  Same checksum as hashFiles would generate, but actually supports env vars.

Also stop trying to use previous job output as an env var in subsequent jobs, because apparently that context also isn't available in jobs.env.  The value is only used once per later job, so just directly use actions substitution instead of a more-readable env var.  Siiiigggghhhhhh.

And fix all the other little issues so it'll work